+++ examples/anchor/marinade/marinade_finance_client_consumer/src/lib.rs
@@ -1,0 +1,1 @@
+//! All code in tests