+++ src/write_cargotoml.rs
@@ -40,6 +40,7 @@ impl<'a> CargoToml<'a> {
             dependencies: GeneratedCrateDependencies {
                 borsh: DependencyValue(&args.borsh_vers),
                 solana_program: DependencyValue(&args.solana_program_vers),
+                serde: OptionalDependencyValue(&args.serde_vers),
                 thiserror: thiserror.map(DependencyValue),
                 num_derive: num_derive.map(DependencyValue),
                 num_traits: num_traits.map(DependencyValue),
@@ -62,6 +63,8 @@ struct GeneratedCrateDependencies<'a> {
     #[serde(rename = "solana-program")]
     solana_program: DependencyValue<'a>,
 
+    serde: OptionalDependencyValue<'a>,
+
     thiserror: Option<DependencyValue<'a>>,
 
     #[serde(rename = "num-derive")]
@@ -82,5 +85,25 @@ impl Serialize for DependencyValue<'_> {
             Ok(v) => v.serialize(serializer), // "workspace = true"
             Err(_) => self.0.serialize(serializer),
         }
+    }
+}
+
+struct OptionalDependencyValue<'a>(&'a str);
+
+impl Serialize for OptionalDependencyValue<'_> {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        let mut map: Map<_, _> = match toml::from_str(self.0) {
+            Ok(v) => v,
+            Err(_) => {
+                let mut m = Map::new();
+                m.insert("version".to_owned(), self.0.into());
+                m
+            }
+        };
+        map.insert("optional".to_owned(), true.into());
+        map.serialize(serializer)
     }
 }